libft

+++ libft/ft_strlen.c
@@ -3,19 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: obalaban <obalaban@student.42.fr>          +#+  +:+       +#+        */
+/*   By: obalaban <obalaban@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/22 19:32:26 by obalaban          #+#    #+#             */
-/*   Updated: 2022/06/29 18:39:51 by obalaban         ###   ########.fr       */
+/*   Updated: 2022/07/02 14:42:45 by obalaban         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
-#include <stdio.h>
+#include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (s[index])
